+++ CHANGELOG.md
@@ -38,6 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Features
+
+- [#53](https://github.com/alleslabs/celatone-frontend/pull/53) Show contract description in contract details page
 - [#58](https://github.com/alleslabs/celatone-frontend/pull/58) Wireup top section in contract details page 
 - [#54](https://github.com/alleslabs/celatone-frontend/pull/54) Render execute cmds shortcut in contract details page
 - [#46](https://github.com/alleslabs/celatone-frontend/pull/46) Wireup instantiate info in contract details page